module3-task1: * move mock helpers
