+++ src/helpers/helpers.ts
@@ -1,3 +1,2 @@
 export * from './array/array';
 export * from './number/number';
-export * from './mocks/mocks';
