+++ src/mocks/helpers/generate-artist-question-answer.helper.ts
