+++ src/mocks/helpers/generate-artist-question.helper.ts
@@ -1,8 +1,8 @@
 import { getRandomItem, getShuffledArray } from '~/helpers/array/array';
-import { generateArtistQuestionSong } from '~/helpers/mocks/generate-artist-question-song.helper';
-import { generateArtistQuestionAnswer } from '~/helpers/mocks/generate-artist-question-answer.helper';
-import { generateArtistQuestionAnswers } from '~/helpers/mocks/generate-artist-question-answers.helper';
-import { getUniqueAnswerTypes } from '~/helpers/mocks/get-unique-answer-types.helper';
+import { generateArtistQuestionSong } from '~/mocks/helpers/generate-artist-question-song.helper';
+import { generateArtistQuestionAnswer } from '~/mocks/helpers/generate-artist-question-answer.helper';
+import { generateArtistQuestionAnswers } from '~/mocks/helpers/generate-artist-question-answers.helper';
+import { getUniqueAnswerTypes } from '~/mocks/helpers/get-unique-answer-types.helper';
 import { IArtistQuestion } from '~/common/interfaces/interfaces';
 import { QuestionType } from '~/common/enums/enums';
