+++ src/mocks/helpers/generate-genre-question.helper.ts
@@ -1,11 +1,19 @@
 import { getRandomItem, getShuffledArray } from '~/helpers/array/array';
-import { generateGenreQuestionAnswer } from '~/helpers/mocks/generate-genre-question-answer.helper';
-import { generateGenreQuestionAnswers } from '~/helpers/mocks/generate-genre-question-answers.helper';
-import { getUniqueAnswerTypes } from '~/helpers/mocks/get-unique-answer-types.helper';
+import { generateGenreQuestionAnswer } from '~/mocks/helpers/generate-genre-question-answer.helper';
+import { generateGenreQuestionAnswers } from '~/mocks/helpers/generate-genre-question-answers.helper';
+import { getUniqueAnswerTypes } from '~/mocks/helpers/get-unique-answer-types.helper';
 import { IGenreQuestion } from '~/common/interfaces/interfaces';
 import { QuestionType } from '~/common/enums/enums';
 
-const GENRE_TYPES = [`rock`, `blues`, `jazz`, `metal`, `pop`, `country`, `techno`];
+const GENRE_TYPES = [
+  `rock`,
+  `blues`,
+  `jazz`,
+  `metal`,
+  `pop`,
+  `country`,
+  `techno`,
+];
 const CUT_GENRE_START_IDX = 0;
 const ANSWERS_COUNT = 3;
