+++ src/mocks/helpers/get-unique-answer-types.helper.ts
